# apidog/elloot-api.openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Elloot API",
    "description": "Marketplace API (escrow). Import this file in Apidog: Import → OpenAPI.\n\nAuth: use `Authorization: Bearer <accessToken>` from register/login.\nSandbox flow: create order → checkout → POST /payments/sandbox/confirm → deliver → confirm.",
    "version": "0.1.0"
  },
  "servers": [
    {
      "url": "http://localhost:5000",
      "description": "Local (.env.local PORT=5000)"
    },
    {
      "url": "http://localhost:3333",
      "description": "Local default (.env.example)"
    }
  ],
...
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT accessToken from /api/auth/register or /api/auth/login"
      }
    },
    "schemas": {
...
      "CreateListingBody": {
        "type": "object",
        "required": ["categoryId", "title", "description", "priceCents"],
        "properties": {
          "categoryId": { "type": "string" },
          "title": { "type": "string", "minLength": 5, "maxLength": 120 },
          "description": {
            "type": "string",
            "minLength": 20,
            "maxLength": 5000
          },
          "priceCents": {
            "type": "integer",
            "minimum": 1,
            "maximum": 50000000,
            "example": 5000
          },
          "mediaUrls": {
            "type": "array",
            "maxItems": 8,
            "items": { "type": "string", "format": "uri" },
            "default": []
          },
          "publish": {
            "type": "boolean",
            "default": false,
            "description": "If true, listing is created as ACTIVE"
          }
        }
      },
      "UpdateListingBody": {
        "type": "object",
        "minProperties": 1,
        "properties": {
          "categoryId": { "type": "string" },
          "title": { "type": "string", "minLength": 5, "maxLength": 120 },
          "description": {
            "type": "string",
            "minLength": 20,
            "maxLength": 5000
          },
          "priceCents": {
            "type": "integer",
            "minimum": 1,
            "maximum": 50000000
          },
          "mediaUrls": {
            "type": "array",
            "maxItems": 8,
            "items": { "type": "string", "format": "uri" }
          }
        }
      },
...
    }
  },
  "paths": {
...
    "/api/catalog/listings": {
      "get": {
        "tags": ["Catalog"],
        "summary": "Public ACTIVE listings (filters + cursor)",
        "operationId": "listCatalogListings",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "schema": { "type": "integer", "default": 20, "maximum": 50 }
          },
          {
            "name": "game",
            "in": "query",
            "description": "Game id or slug",
            "schema": { "type": "string" }
          },
          {
            "name": "category",
            "in": "query",
            "description": "Category id or slug",
            "schema": { "type": "string" }
          },
          {
            "name": "q",
            "in": "query",
            "schema": { "type": "string" }
          },
          {
            "name": "minPriceCents",
            "in": "query",
            "schema": { "type": "integer" }
          },
          {
            "name": "maxPriceCents",
            "in": "query",
            "schema": { "type": "integer" }
          },
          {
            "name": "cursor",
            "in": "query",
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": { "description": "Listings + nextCursor" }
        }
      }
    },
    "/api/listings": {
      "post": {
        "tags": ["Listings"],
        "summary": "Create listing (promotes BUYER → SELLER)",
        "operationId": "createListing",
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/CreateListingBody" },
              "example": {
                "categoryId": "cmsXXXX",
                "title": "Conta VIP Rank Diamante",
                "description": "Conta completa com skins e rank diamante, entrega imediata após pagamento.",
                "priceCents": 5000,
                "mediaUrls": [],
                "publish": true
              }
            }
          }
        },
        "responses": {
          "201": { "description": "Listing created" },
          "404": { "description": "Category not found" }
        }
      }
    },
...
    "/api/listings/{id}": {
      "get": {
        "tags": ["Listings"],
        "summary": "Get listing by id",
        "operationId": "getListing",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": { "description": "Listing" },
          "404": { "description": "Not found" }
        }
      },
      "patch": {
        "tags": ["Listings"],
        "summary": "Update own listing",
        "operationId": "updateListing",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/UpdateListingBody" }
            }
          }
        },
        "responses": {
          "200": { "description": "Updated" },
          "403": { "description": "Forbidden" },
          "409": { "description": "Already sold" }
        }
      },
      "delete": {
        "tags": ["Listings"],
        "summary": "Soft-delete listing (REMOVED)",
        "operationId": "deleteListing",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": { "description": "Marked REMOVED" }
        }
      }
    },
...
    "/api/disputes": {
      "post": {
        "tags": ["Disputes"],
        "summary": "Open dispute on PAID/DELIVERED order",
        "operationId": "openDispute",
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["orderId", "reason"],
                "properties": {
                  "orderId": { "type": "string" },
                  "reason": { "type": "string", "minLength": 10 }
                }
              }
            }
          }
        },
        "responses": { "201": { "description": "Opened" } }
      }
    },
    "/api/disputes/mine": {
      "get": {
        "tags": ["Disputes"],
        "summary": "My disputes",
        "operationId": "listMyDisputes",
        "security": [{ "bearerAuth": [] }],
        "responses": { "200": { "description": "Disputes" } }
      }
    },
    "/api/disputes/{id}": {
      "get": {
        "tags": ["Disputes"],
        "summary": "Get dispute",
        "operationId": "getDispute",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "responses": { "200": { "description": "Dispute" } }
      }
    },
    "/api/disputes/{id}/resolve": {
      "post": {
        "tags": ["Disputes"],
        "summary": "Resolve dispute (ADMIN)",
        "operationId": "resolveDispute",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["resolution"],
                "properties": {
                  "resolution": {
                    "type": "string",
                    "enum": ["RELEASE_TO_SELLER", "REFUND_BUYER", "PARTIAL"]
                  },
                  "notes": { "type": "string" },
                  "sellerAmountCents": {
                    "type": "integer",
                    "description": "Required for PARTIAL"
                  }
                }
              }
            }
          }
        },
        "responses": { "200": { "description": "Resolved" } }
      }
    },
...
  }
}